task5/django-apache/requirements.txt
asgiref==3.5.2
astroid==2.12.12
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
dill==0.3.5.1
Django==4.1.2
djangorestframework==3.14.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==21.3
platformdirs==2.5.2
psycopg2-binary==2.9.4
pylint==2.15.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytz==2022.5
redis==4.3.4
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
typing-extensions==4.4.0
wrapt==1.14.1
